requirements.txt
python-dotenv
google-generativeai
IPython
streamlit
streamlit python-socketio eventlet